Emulate prior v3 reportingRound in feeds
- It was removed from the contract to save gas
