+++ feeds/src/contracts/FluxAggregatorAbi.json
@@ -817,19 +817,6 @@
   },
   {
     "inputs": [],
-    "name": "reportingRound",
-    "outputs": [
-      {
-        "internalType": "uint256",
-        "name": "",
-        "type": "uint256"
-      }
-    ],
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [],
     "name": "requestNewRound",
     "outputs": [],
     "stateMutability": "nonpayable",
